Section4-ContinuousTraining-Airflow-Composer/assignment/ci-cd/requirements.txt
scikit-learn==1.2.2
pandas
pyarrow
pytest
requests
google-cloud-storage
google-cloud-bigquery